[Task #72273] Sửa các tồn đọng ngày 15/1/2024

+++ app/Services/PostService.php
@@ -77,9 +77,9 @@ class PostService
             $arrHashtagCustom = $this->hashtagRepository->insertCustomPostHashtag($hashtagCustom);
         }
 
-        if (!empty($hashtags)) {
-            $this->postHashtagRepository->insertPostHashtag($postId, $hashtags, $action, $arrHashtagCustom ?? []);
-        } else {
+        $this->postHashtagRepository->insertPostHashtag($postId, $hashtags, $action, $arrHashtagCustom ?? []);
+
+        if (empty($hashtags) && empty($hashtagCustom)) {
             $this->postHashtagRepository->removePostHashtag($postId);
         }
     }
